refactored the register page
