+++ frontend/src/pages/register/components/registerForm/index.tsx
@@ -1,0 +1,104 @@
+import React, { useState, ChangeEvent } from 'react';
+
+import { Box, TextInput, Text } from '@primer/react';
+import {
+  PersonIcon,
+  LockIcon,
+  MailIcon,
+  EyeIcon,
+  EyeClosedIcon,
+} from '@primer/octicons-react';
+
+interface FormData {
+  username: string;
+  password: string;
+  email: string;
+}
+
+interface Props {
+  setFormData: React.Dispatch<React.SetStateAction<FormData>>;
+  formData: FormData;
+}
+
+export const RegisterForm = (props: Props) => {
+  const { setFormData, formData } = props;
+
+  const [passwordVisible, setPasswordVisible] = useState<boolean>(false);
+
+  const handleInputChange = (event: ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = event.target;
+    setFormData((prevData) => ({
+      ...prevData,
+      [name]: value,
+    }));
+  };
+
+  const togglePasswordVisibility = () => {
+    setPasswordVisible((prev) => !prev);
+  };
+
+  const inputStyle = {
+    width: '300px',
+  };
+
+  return (
+    <>
+      {/* Username input */}
+      <Box sx={{ display: 'grid' }}>
+        <Text sx={{ color: 'fg.muted' }}>Username:</Text>
+        <TextInput
+          aria-label="Username"
+          name="username"
+          size="large"
+          placeholder="Username"
+          value={formData.username}
+          onChange={handleInputChange}
+          leadingVisual={PersonIcon}
+          sx={inputStyle}
+          required
+        />
+      </Box>
+
+      {/* Email input */}
+      <Box sx={{ display: 'grid' }}>
+        <Text sx={{ color: 'fg.muted' }}>Email:</Text>
+        <TextInput
+          aria-label="Email"
+          name="email"
+          size="large"
+          placeholder="Email"
+          type="email"
+          value={formData.email}
+          onChange={handleInputChange}
+          leadingVisual={MailIcon}
+          sx={inputStyle}
+          required
+        />
+      </Box>
+
+      {/* Password input */}
+      <Box sx={{ display: 'grid' }}>
+        <Text sx={{ color: 'fg.muted' }}>Password:</Text>
+        <TextInput
+          aria-label="Password"
+          name="password"
+          type={passwordVisible ? 'text' : 'password'}
+          size="large"
+          placeholder="Password"
+          value={formData.password}
+          onChange={handleInputChange}
+          leadingVisual={LockIcon}
+          trailingAction={
+            <TextInput.Action
+              onClick={togglePasswordVisibility}
+              icon={passwordVisible ? EyeClosedIcon : EyeIcon}
+              aria-label={passwordVisible ? 'Hide password' : 'Show password'}
+            />
+          }
+          sx={inputStyle}
+          required
+        />
+      </Box>
+    </>
+  );
+};
