+++ frontend/src/pages/register/index.tsx
@@ -1,116 +1,13 @@
-import React, { useState, ChangeEvent, FormEvent } from 'react';
+import React from 'react';
 
-import { capitalize } from '@utils/capitalize';
-import { fetchWrapper } from '@utils/fetchWrapper';
+import { Box, Button, Text, Heading, Link, Spinner } from '@primer/react';
 
-import {
-  Box,
-  TextInput,
-  Button,
-  Text,
-  Heading,
-  Link,
-  Spinner,
-} from '@primer/react';
-import {
-  PersonIcon,
-  LockIcon,
-  MailIcon,
-  EyeIcon,
-  EyeClosedIcon,
-} from '@primer/octicons-react';
-
-interface FormData {
-  username: string;
-  password: string;
-  email: string;
-}
+import { RegisterForm } from '@pages/register/components/registerForm';
+import { useFormSubmit } from '@pages/register/hooks/useFormSubmit';
 
 export const Register: React.FC = () => {
-  const [passwordVisible, setPasswordVisible] = useState<boolean>(false);
-  const [errorMessage, setErrorMessage] = useState<string>('');
-  const [loading, setLoading] = useState<boolean>(false);
-  const [formData, setFormData] = useState<FormData>({
-    username: '',
-    password: '',
-    email: '',
-  });
-
-  const handleInputChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = event.target;
-    setFormData((prevData) => ({
-      ...prevData,
-      [name]: value,
-    }));
-  };
-
-  const togglePasswordVisibility = () => {
-    setPasswordVisible((prev) => !prev);
-  };
-
-  const validateForm = (): boolean => {
-    const { username, password, email } = formData;
-
-    if (
-      username.trim() === '' ||
-      password.trim() === '' ||
-      email.trim() === ''
-    ) {
-      setErrorMessage('All fields must be filled.');
-      return false;
-    }
-
-    if (password.length < 8) {
-      setErrorMessage('Password must be at least 8 characters long.');
-      return false;
-    }
-
-    // Simple email regex for validation
-    const emailRegex = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
-    if (!emailRegex.test(email)) {
-      setErrorMessage('Invalid email address.');
-      return false;
-    }
-
-    setErrorMessage('');
-    return true;
-  };
-
-  const handleSubmit = async (event: FormEvent) => {
-    event.preventDefault();
-
-    if (!validateForm()) return;
-
-    setLoading(true);
-    try {
-      const backendUrl = import.meta.env.VITE_BACKEND_URL;
-      const response = await fetchWrapper(`${backendUrl}/user/sign-up`, {
-        method: 'POST',
-        body: JSON.stringify(formData),
-      });
-
-      if (response.status !== 200) {
-        setErrorMessage(capitalize(response.response.message) + '.');
-        return;
-      }
-
-      window.location.href = '/sign-in';
-      setFormData({
-        username: '',
-        password: '',
-        email: '',
-      });
-    } catch (err: any) {
-      console.error('Registration error:', err);
-      setErrorMessage(capitalize(err.message || 'An error occurred') + '.');
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const inputStyle = {
-    width: '300px',
-  };
+  const { errorMessage, loading, setFormData, formData, handleSubmit } =
+    useFormSubmit();
 
   return (
     <Box
@@ -131,6 +28,7 @@ export const Register: React.FC = () => {
           p: 4,
         }}
       >
+        {/* Title */}
         <Heading
           sx={{
             fontSize: '28px',
@@ -140,70 +38,22 @@ export const Register: React.FC = () => {
           Sign Up
         </Heading>
 
-        <Box sx={{ display: 'grid' }}>
-          <Text sx={{ color: 'fg.muted' }}>Username:</Text>
-          <TextInput
-            aria-label="Username"
-            name="username"
-            size="large"
-            placeholder="Username"
-            value={formData.username}
-            onChange={handleInputChange}
-            leadingVisual={PersonIcon}
-            sx={inputStyle}
-            required
-          />
-        </Box>
+        {/* Register form */}
+        <RegisterForm setFormData={setFormData} formData={formData} />
 
-        <Box sx={{ display: 'grid' }}>
-          <Text sx={{ color: 'fg.muted' }}>Email:</Text>
-          <TextInput
-            aria-label="Email"
-            name="email"
-            size="large"
-            placeholder="Email"
-            type="email"
-            value={formData.email}
-            onChange={handleInputChange}
-            leadingVisual={MailIcon}
-            sx={inputStyle}
-            required
-          />
-        </Box>
-
-        <Box sx={{ display: 'grid' }}>
-          <Text sx={{ color: 'fg.muted' }}>Password:</Text>
-          <TextInput
-            aria-label="Password"
-            name="password"
-            type={passwordVisible ? 'text' : 'password'}
-            size="large"
-            placeholder="Password"
-            value={formData.password}
-            onChange={handleInputChange}
-            leadingVisual={LockIcon}
-            trailingAction={
-              <TextInput.Action
-                onClick={togglePasswordVisibility}
-                icon={passwordVisible ? EyeClosedIcon : EyeIcon}
-                aria-label={passwordVisible ? 'Hide password' : 'Show password'}
-              />
-            }
-            sx={inputStyle}
-            required
-          />
-        </Box>
-
+        {/* Redirect to sign in */}
         <Text sx={{ textAlign: 'center' }}>
           Have an account? • <Link href="/sign-in">Sign In</Link>
         </Text>
 
+        {/* Error message */}
         {errorMessage && (
           <Text sx={{ color: 'red', fontSize: '14px', textAlign: 'center' }}>
             {errorMessage}
           </Text>
         )}
 
+        {/* Submit Button  */}
         <Button
           type="submit"
           disabled={loading}
